% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioDataScience-package.R
\docType{package}
\name{BioDataScience-package}
\alias{BioDataScience-package}
\title{Base configuration for BioDataScience}
\description{
Configure the R environment for the BioDartaScience course using \code{\link[=init]{init()}}.
}
