% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biometrie.R
\docType{data}
\name{biometrie}
\alias{biometrie}
\title{Mesure de biometrie humaine}
\format{An object of class \code{dataframe} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 396 rows and 8 columns.}
\usage{
biometrie
}
\description{
Compilation d'un ensemble de mesures de biometrie réalisée par les étudiants de l'UMONS dans le cadre de biostatistique et probabilités entre 2013 et 2017.
}
\examples{
data(biometrie)
class(biometrie)
head(biometrie)
plot(biometrie)
}
\keyword{datasets}
