% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biometry}
\alias{biometry}
\title{Measure of human biometry at Hainaut, Belgium}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 395 rows and 7 columns.}
\usage{
biometry
}
\description{
Measure of human biometry to study health status with several index
as Body Weight Index (BMI) between 2013 an 2017.
}
\keyword{datasets}
