% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_login.R
\name{check_login}
\alias{check_login}
\title{Check login, Internet and database access}
\usage{
check_login(login, email)
}
\arguments{
\item{login}{The user's login (same one as Github login).}

\item{email}{The user's email address.}
}
\value{
A vector of three boolean for \code{internet}, \code{login} and \code{db} checks,
plus a comment with a message.
}
\description{
This function checks if the login and email provided are already known in
the SDD database. It also checks Internet access and SDD database access.
}
\seealso{
\code{\link[=run]{run()}}, \code{\link[=record_sdd]{record_sdd()}}
}
\concept{check login for the BioDataScience package}
\keyword{utilities}
