% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker_sdd.R
\name{checker_sdd}
\alias{checker_sdd}
\title{A default checker that just aknowledge submission}
\usage{
checker_sdd(
  label,
  user_code,
  solution_code,
  check_code,
  envir_result,
  evaluate_result,
  ...
)
}
\arguments{
\item{label}{The label for the learnr exercice.}

\item{user_code}{The code submitted by the user.}

\item{solution_code}{The code provided by the "-solution" chunk.}

\item{check_code}{The code provided by the "-check" chunk.}

\item{envir_result}{The environment after the execution of the chunk.}

\item{evaluate_result}{Result from evaluation of the code.}

\item{...}{Additional parameters (currently not used).}
}
\value{
A list with components \code{message}, \code{correct} and \code{location}.
}
\description{
Check code submitted during an exercice. This version just
acknowledges reception of the submission. This function is used internally
by the tutorials and is not intended for the end-user.
}
\seealso{
\code{\link[=run]{run()}}
}
\concept{record events from the BioDataScience package}
\keyword{utilities}
