% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_rmd.R
\name{corr_rmd}
\alias{corr_rmd}
\title{Convert to HTML documents all Rmarkdown in a project}
\usage{
corr_rmd(path = ".", correction = TRUE)
}
\arguments{
\item{path}{the path of the project}

\item{correction}{If correction is TRUE, a copy of rmd file do not convert is created.}
}
\value{
a data.frame with the results of the compilations in html
}
\description{
Find all R markdown files in a Rstudio project
and try to convert then an HTML document with \code{render_sdd()}.
}
\examples{
#library(pfunctions)
#test <- corr_project()

}
