% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk_info.R
\name{disk_info}
\alias{disk_info}
\title{Show disk information}
\usage{
disk_info(path = "/home/jovyan", intern = FALSE)
}
\arguments{
\item{path}{The path to look for.}

\item{intern}{If \code{TRUE}, the output is returned as a character vector,
otherwise the output is directly printed at the console.}
}
\value{
With \code{intern = TRUE}, a character vector with the output of the
command. Otherwise, the function is used for its side effect of printing the
information at the console.
}
\description{
Calculate disk size total, used and free for a path (or for all disks if
\code{path = ""}) using the \code{df} unix command.
}
\examples{
disk_info()
disk_info("")
}
