% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{.get_cursor_col}
\alias{.get_cursor_col}
\title{get the column the cursor is on from a selection}
\usage{
.get_cursor_col(current_selection)
}
\arguments{
\item{current_selection}{a selection from the rstudioapi document context}
}
\description{
get the column the cursor is on from a selection
}
