% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{.get_cursor_line}
\alias{.get_cursor_line}
\title{get the line the cursor is on}
\usage{
.get_cursor_line(
  context,
  current_selection = rstudioapi::primary_selection(context)
)
}
\arguments{
\item{context}{the rtsudioapi document context}

\item{current_selection}{the selection to find the line for, defaults to
primary selection}
}
\description{
get the line the cursor is on
}
