% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{.sdd_help_addin}
\alias{.sdd_help_addin}
\title{Get help for the BioDataScience course}
\usage{
.sdd_help_addin()
}
\value{
The corresponding R code for the command that is invoked is returned
invisibly as a noquote object.
}
\description{
This fonction uses a Shiny gadget and RStudio context. It is only available
from the addins menu in RStudio.
}
