% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize the environment for our BioDataScience courses}
\usage{
init()
}
\value{
Nothing, the function is used for its side-effect of creating global
variables.
}
\description{
This function creates several environment variables that are widely used by
other packages and applications. Currently, \code{MONGO_URL}, \code{MONGO_BASE},
\code{MONGO_USER} and \code{MONGO_PASSWORD} are set to default values if not already
there.
}
\examples{
# Use this at the beginning of your learndown Shiny or learnr applications
BioDataScience::init()
# Now, you have access to MONGO_URL and MONGO_BASE
Sys.getenv("MONGO_URL")
Sys.getenv("MONGO_BASE")
}
