% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_all_rmd_project.R
\name{render_all_rmd_project}
\alias{render_all_rmd_project}
\title{Convert to HTML documents all Rmarkdown in a project}
\usage{
render_all_rmd_project(path = ".", correction = FALSE)
}
\arguments{
\item{path}{the path of the project}

\item{correction}{If the correction is TRUE, a copy of rmd file do not convert is created.}
}
\value{
a data.frame with the results of the compilations in html
}
\description{
Find all R markdown files in a Rstudio project and try to convert then an HTML document.
This functions use \code{html_document()} and a serie of predefined arguments
in order to meet the requirements of the biological data science course.
}
\examples{
#library(pfunctions)
#test <- render_all_rmd_project()

}
