% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_all_rmd_project.R
\name{render_sdd}
\alias{render_sdd}
\title{Convert to HTML documents an Rmarkdown with specific rules}
\usage{
render_sdd(path, ...)
}
\arguments{
\item{path}{the path of the Rmd file.}

\item{...}{The arguments of render()}
}
\value{
an html document
}
\description{
Convert Rmarkdown file in an HTML document a serie of predefined arguments
in order to meet the requirements of the biological data science course.
This functions use \code{render()} and \code{html_document()}
}
\examples{
#library(pfunctions)
#test <- render_sdd(path = "file.Rmd")

}
