% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run learnr tutorials from the BioDataScience package}
\usage{
run(tutorial, ...)
}
\arguments{
\item{tutorial}{The name of the tutorial to use. If not provided, a list of
available tutorials is displayed.}

\item{...}{Further arguments passed to \code{\link[=run_tutorial]{run_tutorial()}}}
}
\value{
If \code{tutorial} is not provided, a list of available tutorials.
}
\description{
Start the learnr R engine in the current R session with the
selected tutorial.
}
\examples{
# To list the availalble tutorials:
run()
\dontrun{
run("module02a_nuage_de_points")
}
}
\seealso{
\code{\link[=run_tutorial]{run_tutorial()}}
}
\concept{run interactive learnr documents from the BioDataScience package}
\keyword{utilities}
