% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_mail_sdd.R
\name{send_mail_sdd}
\alias{send_mail_sdd}
\title{Send your BioDataScience submissions by email}
\usage{
send_mail_sdd(file = "~/.local/share/R/learnr/biodatascience")
}
\arguments{
\item{file}{The file that contains your submissions information.}
}
\value{
The data are returned invisibly.
}
\description{
Your submissions are send to a central database. However, in
case that database is not accessible, the data is stored locally. This
function uses your plain email to send your records. Note that, once the
email is created, the local version of your records is reset. So, if you
finally decide to NOT send the eamil, these records are lost (in this case,
call your teachers to recover them, if you have to.)
}
\examples{
# To start from a list of available tutorials:
run()
\dontrun{
run("02b_nuage_points")
}
}
\seealso{
\code{\link[=run]{run()}}
}
\concept{run interactive learnr documents from the BioDataScience package}
\keyword{utilities}
