% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_repo.R
\name{switch_repo}
\alias{switch_repo}
\title{Switch the remote URL of a GitHUb repository between HTTPS and SSH}
\usage{
switch_repo(verbose = TRUE)
}
\arguments{
\item{verbose}{If \code{TRUE} (by default), the function produces a message that explain what
was done.}
}
\value{
The new remote URL for the GitHub repository is returned.
}
\description{
Use this function to easily switch the remote URL of a GitHub repository
between HTTPS and SSH. GitHub allows for the two modes, but if the machine is
configured in one mode and a clone is created in the other mode, one cannot
push or pull. The correction is easy to do at the command line, but this
function makes it even easier.
}
\examples{
# Run only when you know what you are doing: the remote URL is changed!
#switch_repo()
}
