% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioDataScience-package.R
\docType{package}
\name{BioDataScience-package}
\alias{BioDataScience}
\alias{BioDataScience-package}
\title{Base configuration functions for BioDataScience course at the University of Mons, Belgium}
\description{
Record user informations using \code{\link[=sign_in]{sign_in()}} (or erase them using \code{\link[=sign_out]{sign_out()}}).
Access the MongoDB database for the course through \code{\link[=config]{config()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BioDataScience-Course/BioDataScience}
  \item Report bugs at \url{https://github.com/BioDataScience-Course/BioDataScience/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

Authors:
\itemize{
  \item Guyliann Engels \email{guyliann.engels@umons.ac.be} (\href{https://orcid.org/0000-0001-9514-1014}{ORCID})
}

}
