% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biometrie}
\alias{biometrie}
\title{Mesure de biometrie humaine}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 395 rows and 7 columns.}
\usage{
biometrie
}
\description{
Compilation d'un ensemble de mesures biometriques réalisée par les étudiants
de l'UMONS dans le cadre du cours de biostatistique et probabilités
entre 2013 et 2017.
}
\examples{
data(biometrie)
class(biometrie)
head(biometrie)
plot(biometrie)
}
\keyword{datasets}
