% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biometry}
\alias{biometry}
\title{Measure of human biometry at Hainaut, Belgium}
\format{
A data frame with 7 variables and 395 observations:
\describe{
\item{\code{gender}}{A \strong{factor} with two levels: \code{"Men"}, and \code{"Women"}.}
\item{\code{day_birth}}{Day of birth (YYYY-MM-DD).}
\item{\code{weight}}{Weight (in kg).}
\item{\code{height}}{Height (in cm).}
\item{\code{wrist}}{Wrist circumfernce (in mm).}
\item{\code{year_measure}}{Year of the biometric measure.}
\item{\code{age}}{Age of people.}
}
}
\usage{
biometry
}
\description{
Measure of human biometry to study health status with several index
as Body Weight Index (BMI) between 2013 an 2017.
}
\keyword{datasets}
