% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize the environment for our BioDataScience courses}
\usage{
init()
}
\value{
Nothing, the function is used for its side-effect of creating global
variables.
}
\description{
This function creates several environment variables that are used by other
packages and applications. Currently, \code{MONGO_URL}, \code{MONGO_URL_SERVER}
(the later one is used if the application is run from a server),
\code{MONGO_BASE}, \code{MONGO_USER}, \code{MONGO_PASSWORD} and \code{LOCAL_STORAGE} are set to
default values if not already defined.
}
\examples{
# Use this at the beginning of your learndown Shiny or learnr applications
BioDataScience::init()
# Now, you have access to MONGO_URL and MONGO_BASE and others
Sys.getenv("MONGO_URL")
Sys.getenv("MONGO_BASE")
}
