% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize the environment for our BioDataScience courses}
\usage{
init()
}
\value{
Nothing, the function is used for its side-effect of creating global
variables.
}
\description{
Initialize the environment for our BioDataScience courses
}
\examples{
# Use this at the beginning of your learndown Shiny or learnr applications
BioDataScience::init()
# Now, you have access to MONGO_URL and MONGO_BASE
Sys.getenv("MONGO_URL")
Sys.getenv("MONGO_BASE")
}
